Fix drag and drop for sorted todo and done groups
