+++ src/TodoList.jsx
@@ -1,50 +1,116 @@
-import React from "react";
+import React, { useState } from "react";
 import styled from "@emotion/styled";
 import PropTypes from "prop-types";
 import { observer } from 'mobx-react-lite';
 import TodoItem from "./TodoItem";
 import todoStore from "./ObservableTodoStore";
 
+  // Following drag and drop tutorial at
+  // https://dev.to/florantara/creating-a-drag-and-drop-list-with-react-hooks-4c0i
+  const initialDnDState = {
+    draggedFrom: null,
+    draggedTo: null,
+    isDragging: false,
+    originalOrder: [],
+    updatedOrder: [],
+  }
+
 const UnstyledTodoList = observer(({
   className,
-  handleDragStart,
-  handleDragOver,
-  handleDrop,
 }) => {
   const {
-    completedTodos,
+    sortedTodos,
     completedTodosCount,
-    incompletedTodos,
     incompletedTodosCount,
+    totalTodosCount,
    } = todoStore;
 
-  const incompletedTodoItems = incompletedTodos.map((todoItem, todoIndex) => {
+  const [dragAndDrop, setDragAndDrop] = useState(initialDnDState);
+
+  const handleDragStart = (dragEvent) => {
+    // Access the "data-position" attr of the current element being dragged
+    const initialPosition = Number(dragEvent.currentTarget.dataset.position);
+
+    setDragAndDrop({
+      ...dragAndDrop,
+      draggedFrom: initialPosition,
+      isDragging: true,
+      originalOrder: [...sortedTodos],
+    });
+
+    // Note: This is only needed for Firefox.
+    // Otherwise, DragAndDrop won't work.
+    // But we aren't using it.
+    dragEvent.dataTransfer.setData('text/html', '');
+  };
+
+  const handleDragOver = (dragEvent) => {
+    dragEvent.preventDefault();
+
+    let newTodos = dragAndDrop.originalOrder;
+    const { draggedFrom } = dragAndDrop;
+
+    // Index of the drop area being hovered
+    const draggedTo = Number(dragEvent.currentTarget.dataset.position);
+    // Get the element that's at the position of 'draggedFrom'
+    const todoDragged = newTodos[draggedFrom];
+    const remainingItems = newTodos.filter((item, index) => index !== draggedFrom);
+
+    // Update the list
+    newTodos = [
+      ...remainingItems.slice(0, draggedTo),
+      todoDragged,
+      ...remainingItems.slice(draggedTo),
+    ];
+
+    // Since this event fires many times, check if the targets are different
+    // before updating.
+    if(draggedTo !== dragAndDrop.draggedTo) {
+      setDragAndDrop({...dragAndDrop,
+      updatedOrder: newTodos,
+      draggedTo,
+    })
+    }
+  };
+
+  const handleDrop = () => {
+    todoStore.setTodos(dragAndDrop.updatedOrder);
+
+    // Reset the state of the drag and drop
+    setDragAndDrop({
+      ...dragAndDrop,
+      draggedFrom: null,
+      draggedTo: null,
+      isDragging: false,
+    });
+  }
+
+  const sortedTodoData = sortedTodos.map((todoItem, todoIndex) => {
     const todoKey = `${todoIndex}`;
-    return (
-      <TodoItem
-        key={todoKey}
-        index={todoIndex}
-        todo={todoItem}
-        handleDragStart={handleDragStart}
-        handleDragOver={handleDragOver}
-        handleDrop={handleDrop}
-      />
-    );
+
+    return {
+        completed: todoItem.completed,
+        item: (
+          <TodoItem
+            key={todoKey}
+            index={todoIndex}
+            todo={todoItem}
+            handleDragStart={handleDragStart}
+            handleDragOver={handleDragOver}
+            handleDrop={handleDrop}
+          />
+        ),
+    };
   });
 
-  const completedTodoItems = completedTodos.map((todoItem, todoIndex) => {
-    const todoKey = `${todoIndex}`;
-    return (
-      <TodoItem
-        key={todoKey}
-        index={todoIndex}
-        todo={todoItem}
-        handleDragStart={handleDragStart}
-        handleDragOver={handleDragOver}
-        handleDrop={handleDrop}
-      />
-    );
-  });
+  const incompletedTodoItems  = sortedTodoData
+    .filter(({completed}) => !completed)
+    .map(({item}) => item);
+
+
+  const completedTodoItems  = sortedTodoData
+    .filter(({completed}) => completed)
+    .map(({item}) => item);
 
   const completedEmptyStateMessage = (
     <li>
@@ -79,9 +145,6 @@ const UnstyledTodoList = observer(({
 
 UnstyledTodoList.propTypes = {
   className: PropTypes.string.isRequired,
-  handleDragStart: PropTypes.func.isRequired,
-  handleDragOver: PropTypes.func.isRequired,
-  handleDrop: PropTypes.func.isRequired,
 };
 
 const TodoList = styled(UnstyledTodoList)`
